Increase performance with large amounts of logs
Before being passed into `grep`, the log files are sorted by date. As
`grep` searches over the files, the output is processed. As soon as the
buffer has enough results and `grep` has finished processing logs from
the current date, the process is terminated. The results are then sorted
and truncated as usual.

This decreases `grep`'s resource usage for queries that produce a lot of
results. Queries that only produce a small amount of results will still
require `grep` to search through all the available logs.

Fixes #1