Fix bug with binary file detection
`grep` was detecting some log files as binary and was outputting
`"Binary file [file] matches"` to stdout. This doesn't match the regex
the module uses to process the output and was causing an exception.

This commit fixes this by forcing `grep` to treat all files as text
since there shouldn't be any binary files in the log paths.

It also adds the '--with-filenames' and '-e' switches to guard against a
few other edge cases.